declarative_data_framework/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
[project]
name = "declarative-data-framework"
version = "0.1.0"
description = "A declarative framework for data pipelines."
authors = [
	{ name = "Felipe Sampaio" }
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
	"pyyaml",
	"pyspark",
	"pydantic"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
